# docs/swagger.yaml
basePath: /swagger/things-service
definitions:
...
  common.Page:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  entity.FileUploadReq:
    properties:
      data:
        type: string
      file_name:
        type: string
      mime_type:
        type: string
    type: object
...
  model.Device:
    properties:
      created_by:
        description: 创建人
        type: string
      created_time:
        description: 创建时间
        type: string
      enabled:
        description: 启用禁用
        type: integer
      group_id:
        description: 分组id
        type: string
      id:
        description: 唯一标识(md5计算)
        type: string
      identifier:
        description: 标识
        type: string
      name:
        description: 名称
        type: string
      parent_id:
        description: 父id
        type: string
      product_id:
        description: 产品id
        type: string
      protocol_config:
        description: 协议配置(json)
        type: string
      status:
        description: 状态(0:离线:1在线,2:告警)
        type: integer
      type:
        description: 类型(1设备，2网关)
        type: integer
      updated_by:
        description: 更新人
        type: string
      updated_time:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
  description: things-service API 文档.
  title: things-service RESTful API
  version: "1.0"
paths:
...
  /device:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: parent_id
        in: query
        name: parent_id
        type: string
      - description: group_id
        in: query
        name: group_id
        type: string
      - description: product_id
        in: query
        name: product_id
        type: string
      - description: protocol_config
        in: query
        name: protocol_config
        type: string
      - description: identifier
        in: query
        name: identifier
        type: string
      - description: enabled
        in: query
        name: enabled
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Device'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - Device
    post:
      consumes:
      - application/json
      description: save
      parameters:
      - description: object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Device'
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Device'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: save
      tags:
      - Device
...
  /device-import/excel-import-binary:
    post:
      description: 导入excel,文件字节base64方式
      parameters:
      - description: 文件上传请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/entity.FileUploadReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"status":0,"data":{},"msg":"success"}'
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 导入excel
      tags:
      - 设备导入
  /device-import/excel-modbus-import:
    post:
      description: 导入excel,modbus格式
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: '{"status":0,"data":{},"msg":"success"}'
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 导入excel
      tags:
      - 设备导入
...
  /device/{id}:
    delete:
      description: delete
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Device'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: delete
      tags:
      - Device
  /device/batch-delete:
    post:
      consumes:
      - application/json
      description: batch delete
      parameters:
      - description: id array
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch delete
      tags:
      - Device
  /device/batch-upsert:
    post:
      consumes:
      - application/json
      description: batch update
      parameters:
      - description: objects array
        in: body
        name: entities
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch update
      tags:
      - Device
  /device/groupby:
    get:
      description: GroupBy, for example,  _select=level, then return  {level_val1:sum1,level_val2:sum2},
        _where can input status=0
      parameters:
      - description: _select
        in: query
        name: _select
        required: true
        type: string
      - description: _where
        in: query
        name: _where
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: GroupBy
      tags:
      - Device
  /device/page:
    get:
      description: page query, _page(from 1 begin), _page_size, _order, and others
        fields, status=1, name=$like.%CAM%
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: parent_id
        in: query
        name: parent_id
        type: string
      - description: group_id
        in: query
        name: group_id
        type: string
      - description: product_id
        in: query
        name: product_id
        type: string
      - description: protocol_config
        in: query
        name: protocol_config
        type: string
      - description: identifier
        in: query
        name: identifier
        type: string
      - description: enabled
        in: query
        name: enabled
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.Device'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: page query
      tags:
      - Device
...
swagger: "2.0"